requirements.txt
numpy
scikit_learn
scipy
panda
streamlit